feat: useSnackbar open 로직 변경
