+++ src/shared/lib/index.ts
@@ -1,5 +1,6 @@
 export { useAppDispatch } from './useAppDispatch';
 export { useAppSelector } from './useAppSelector';
+export { useSnackbar, SnackbarContext } from './useSnackbar';
 
 export { fetchAPI } from './fetch';
 export { KSTDate, dateToISO } from './date';
